requirements.txt
flask>=3.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
Pillow>=10.2.0
numpy==1.26.4
pyzbar>=0.1.9
scipy>=1.10.0
pyyaml>=6.0

